+++ packages/material/src/RQBMaterialContext.ts
@@ -2,8 +2,12 @@ import type { Context } from 'react';
 import { createContext } from 'react';
 import type { RQBMaterialComponents } from './types';
 
+export interface RQBMaterialContextValue extends RQBMaterialComponents {
+  showInputLabels?: boolean;
+}
+
 /**
  * @group Components
  */
-export const RQBMaterialContext: Context<RQBMaterialComponents | null> =
-  createContext<RQBMaterialComponents | null>(null);
+export const RQBMaterialContext: Context<RQBMaterialContextValue | null> =
+  createContext<RQBMaterialContextValue | null>(null);