+++ packages/material/src/index.tsx
@@ -14,7 +14,7 @@ import { MaterialValueEditor } from './MaterialValueEditor';
 import { MaterialValueSelector } from './MaterialValueSelector';
 import { RQBMaterialContext } from './RQBMaterialContext';
 import { materialTranslations } from './translations';
-import type { RQBMaterialComponents } from './types';
+import type { MuiAugmentation } from './types';
 import { useMuiComponents } from './useMuiComponents';
 
 export * from './MaterialActionElement';
@@ -48,14 +48,16 @@ const MaterialContextProvider = getCompatContextProvider({
 /**
  * @group Components
  */
-export const QueryBuilderMaterial: QueryBuilderContextProvider<{
-  muiComponents?: RQBMaterialComponents;
-}> = ({ muiComponents: muiComponentsProp, ...props }) => {
+export const QueryBuilderMaterial: QueryBuilderContextProvider<MuiAugmentation> = ({
+  muiComponents: muiComponentsProp,
+  showInputLabels,
+  ...props
+}) => {
   const muiComponents = useMuiComponents(muiComponentsProp);
 
   const ctxValue = useMemo(
-    () => ({ ...muiComponents, ...muiComponentsProp }),
-    [muiComponents, muiComponentsProp]
+    () => ({ ...muiComponents, ...muiComponentsProp, showInputLabels }),
+    [muiComponents, muiComponentsProp, showInputLabels]
   );
 
   return (